Input Handler Setup & Basic Movement
An input handler scriptable object has been set up to allow objects to subscribe functions onto the input being received. A player controller class has been created which binds functions for movement and jumping. A very basic movement component has been put in place for testing purposes, and the jump is yet to be implemented, but prints debug messages to the console to demonstrate the input is working correctly.