Started State Machine
Started a state machine that allows the player to be moved around and jump. This required changes to the previous code, and also made some of the input classes obsolete, so it is yet to be decided if this will be the method used in for the final player controls.