Improved Character Movement & Jump
Improved the feel of character movement by making it feel more responsive, quickly changing opposite directions, and rotating to face the direction the player is moving in.
The jump has been greatly improved, with the 3 portions of a jump: the raise, apex and fall, being split out individually. Mathematical concepts have been applied to each portion of the jump to allow it to be tweaked and customised to get the desired feel for the controls.