Jump Bug Fixes & Input Buffering
Fixed a few bugs with the jump, and split the code into multiple classes to make the implementation cleaner.
Added an input buffer class to create a buffer with a length for any desired input. An input buffer manager class holds references to every type of buffer, and their triggers can be subscribed to the input C# events in the input handler class.
This was used to create a grace timer when the player presses the jump button. If they press it slightly too early before landing, the character will perform the jump when they make contact with the ground. This helps make the controls feel more responsive as the character will perform the actions based off of the players input, even if they are slightly miss-timed.