Coyote Time
Added a short countdown timer whenever the player moves off of a platform and starts falling. If the jump button is pressed before the countdown reaches zero, the character will perform a jump. This helps the player if they press the jump input too late, making the controls feel more responsive.