Deceleration, Movement Code Refactor & Added Spin Attack
Added Deceleration when the player releases the move button. Rather than coming to a complete stand still, the character will gradually slow down until they stop. Like acceleration, an animation curve is provided in the inspector to adjust the smoothness of the transition.
Refactored the movement code, so the different sections of calculations have been separated out into their own functions. A handle movement function calls into all of these in the correct order, calculating the values to be used to set the characters velocity.
The acceleration was changed into a general motion class, which holds variables for an animation curve and velocity change duration, which could then be reused to create the deceleration as well. When calculating the speed change, the object can be passed in to specify whether to acceleration, or deceleration, based on the circumstances.
Expanding on the player state machine, a new state was added where the player can perform a spin attack. If they attack while in mid-air, a slight jump boost is applied to the character. An animation plays to quickly rotate the model, visualising the effect, but a particle effect could also be added to enhance the juiciness of the action. A collider is used to detect any nearby objects while attacking.
Objects that can be broken by being attacked by the player can implement a new 'IBreakable' interface. Now, any class can define how it responds to being hit by the player. More interfaces could be added in the future for more specific behaviour, such as an 'IHitable' for objects that could animate or produce items instead of breaking.