Particles Using VFX Graph
Created a dust trail using the VFX graph for when the player is running around., as well as a ring of clouds for when the player jumps and lands. A player VFX class has been made to store all the particle systems the player will use, and functions to activate, deactivate, or spawn these in. The player state machine can call into this accordingly, triggering and stopping particles as the player enters and leaves different states.