Chest With Coin Spawning
The chest in the prototype level can now be broken open when spin attacked. An animation is played, and a VFX game object is triggered.
 A new coin spawner class has been created as a base abstract class to spawn coins from. This may change to a more generic item spawner class in future, or derive from one as to keep it a more specific type of spawner. From this, a spawn coins circular class was made to spawn a set amount of coins in a circle, within a given radius. These are currently animated using a Vector3 lerp function to transition them from the chest to their position on the circle, but the next thing to implement will be a sin wave to make them 'jump' out, rather than slide.
A basic coin class was made to placed on the coin prefab, but will likely be expanded as more systems are added to the game, such as UI display, score etc...