Crate Fracture & Break VFX
Created a fractured cube 3D model in blender, which was then exported into Unity. When a crate is destroyed, each piece of the fractured mesh has force applied to it, making them launch outwards from the centre of the crate. The pieces gradually shrink in size until they are no longer visible. An additional VFX was added to improve the juiciness of the effect, helping give it a more explosive feel.