Ground Pound
Added a ground pound mechanic, allowing the player to crush objects beneath them, or to be used to come to a quick stop while in the air. This was worked into the state machine, and new VFX have been made to enhance the feel of the action. A shockwave VFX has been left out as it currently doesn't integrate with the other particles, but this can be reworked in future and introduced back in.