High Jump
Added a high jump variation to the normal jump. If the player presses the jump button within a short period of time after landing from a ground pound, they perform a higher type of jump. An animation is played, and a new script subscribes to an event to trigger a trail renderer to create a greater sense of momentum.