Fixed Getting Stuck On Walls & Custom Friction
Fixed an issue where the player would get stuck on walls when moving and jumping. The original solution was to stop receiving movement input when the player was moving directly towards a wall in front of them. This mostly worked, however if the player tried to move at a sideways angle across the wall, the resulting movement stuttered a lot
Instead, physics materials with 0 friction have been applied to all objects in the scene, and a custom friction script can be added to any physics objects to allow the friction and rotational drag to be modified.