Camera Shake, Controller Rumble & Double Spin
Added a camera shake system that has a manager script which holds references to different types of camera shakes (with varying intensities and durations) in the form of scriptable objects. The manager subscribes to events called in the logic of other scripts to respond when certain actions are triggered.
A similar approach has been used for a controller rumble system, however the gamepad currently being used doesn't support rumble through Unity, so this hasn't been fully tested out. yet.
Refactored some of the player code to allow them to perform a double spin while in the air. Once the player has span twice while in the air, they can no longer perform any more spin attacks until they land on the ground, when the counter is reset.