Improved Camera Shake System
used online resources to discover ways of 'juicing' a game camera using mathematical concepts. The structure of the camera shake class has been changed to allow camera shakes to be stacked on top of each other for an increasing effect. The shake also smoothly transitions out, and extra parameters to control the max yaw, pitch, roll and translational movement have been added. Refactored the class to make the code cleaner and easier to setup camera shakes with different events.