Refactored Code
Moved lots of the extra systems (such as animations, particles and camera shakes) into an event manager using the observer design pattern. Now the player logic code doesn't depend on these systems to work, and instead, the event manager subscribes them to the players actions to respond as events are sent out.
If the observers weren't registered to the subject correctly, no errors or bugs appear as the player code can continue to run fine without the need for these systems, however the effects of these systems won't appear in the game.