Jet Pack & Bug Fixes
Implemented a jet pack mechanic, where the player can hold down the jump button while in mid-air to hover for a set duration of time. The duration and initial air boost are parameters that can be customised to adjust the feel of the jet pack. The movement logic may be reworked later to reduce the amount of control the player has while in the air. Fixed some additional bugs with falling and jumping.