Jet Pack VFX
Added a VFX graph particle system to play when the players jet pack is in use. The shader that was used in the original attempt has been left in the project, but no longer included in the scene. Events have been set up to trigger when the VFX should start and stop. Extra bug fixes have been done, and a few quality of life improvements for the input have been addressed to make the controls easier and more accessible to use.