Bug Fixes & Scene Reloading
The scene can now be reloaded when the select button on the gamepad is pressed. The setup for a scene loader, and UI manager have been created, in order to fade the scene out and reload it. There were bugs with the player state machine when scenes are loaded, as the state exit function isn't called so the subscribed events are still active in the next scene, but trying to reference objects that no longer exist. Now whenever a scene is loaded, there is some clean up to ensure the events in the player state machine are unsubscribed.