Improved Hover & Bug Fixes
Added an animation curve to customise the shape of the air boost when performing the hover action. Started work on a hover movement class to reduce the amount of air control the player has, but this will be expanded on, or the normal movement component will be used with it's parameters changed. Fixed more bugs, such as the grounded system still thinking the player is in the air if they spin quick enough after they jump that they're still on the ground.