Dive Mechanic
When the player is ground pounding, if they press the spin attack button, rather than performing a spin, the player will dive forwards. While they gain a good boost in speed, they sacrifice manoeuvrability as they can no longer control the character with the joystick until the dive is finished. The movement may be reviewed later on, to allow the player to have slight control over where they move while diving, but the amount will be very minimal.
A basic event has been set up to trigger a dive animation (to be swapped out later on when official animations have been made), but more events will be subscribed to trigger VFX and a squash and stretch effect.