Improved Dive, Bug Fixes & General Control Tweaks
Improved the Dive feature with extra VFX and animation. Added a movement strength which allows the player to slightly move side to side while diving, or pull the joystick back in the opposite direction to slow the move down. The dot product was used to ensure that no additional speed was added to the dive if the player was tilting the joystick in the same direction that they are already moving in.
Fixed some bugs with the player, such as transitioning into the dive from the ground pound, and prevented some of the VFX particles from being instantiated more than once.
The controls have been play tested and their values tweaked, to adjust how the game feels, but this is subject to change as development continues, more features are implemented, and level / environments are designed.