Major Code Refactoring
After a couple hours researching into the 'SOLID' principles, I've reviewed the code and looked for areas where it can be improved, made more efficient, or adjusted to help it's readability.
The player movement settings have been separated into their own scriptable object, meaning multiple assets with different variables can be made, and the settings that the movement component uses can be swapped out to change how the player moves.
The rotation and velocity calculations of the movement class have been separated out into their own classes. The movement class only focuses on moving the player's rigidbody, while the rotation handles rotating the character towards their move direction, and the velocity class calculates all the values to be used by the movement class.
The player event manager has been re-designed with interfaces, meaning the events no longer need to be set up in the inspector, and will work automatically once programmed.
The ground pound class has been split apart, and is now half the size of what it originally was. The different phases of the ground pound (the build up, falling and landing) have been placed into their own classes, only containing the logic to that specific part of the ground pound. In the main class, a delegate holds a reference to the function it should be calling, so as the ground pound progresses, the function call can be swapped to the logic of next class, meaning only one part of the movement is ever running at one time.