Cinemachine Camera Controller
Setup a Cinemachine 'Look At' camera to follow the player as they move around the level. The camera can be rotated using the mouse or right stick on a gamepad, but it currently seems to jitter after a few seconds into playing the scene, so a solution will need to be found for that. The player controller has been updated to account for the player being able to move in the direction the camera is facing, rather than in world space co-ordinates. The camera shake will need to be re-worked, as the old solution isn't compatible with cinemachine.