Various Mechanic Fixes
Fixed issues with the camera so it now feels nicer to control. Re-integrated the camera shakes to work with cinemachine. Started the grapple / swing mechanic - the player can currently latch onto a test object and hang from it, as well as disconnect to drop back down onto the platform underneath it. Improved the ground pound build up based on feedback, and added a slight air boost to emphasise the build up effect, and make the action feel more powerful.