Swinging Mechanic
The player can now swing around from the object they are attached to.  A line render is shot out to simulate a rope connecting the player to the object, with a spring effect to make it slightly wavy. The player can either drop off or jump from the rope.
A detection system has been added to automatically select the closest target that's closest to the player. The object currently gets highlighted in red to visualise during the prototype phase, but later on a world space UI element will be introduced to indicate to the player which object they are currently targeting.