FOV Effects & Camera Shakes When Swinging
Added a script to change the FOV of the camera. This is triggered when the player either performs a dive or jumps from a swing to help convey a sense of speed. An additional camera shakes has also been added for when the player lands after jumping from a swing.