Coin Updates & UI
Added a coin counter UI to display the current amount of coins the player has. Whenever the player collects a coin, the UI appears on the screen, and if the player hasn't collected any coins after a few seconds, the UI fades out. The coins are now sucked towards the player as they get close enough to them, which prevents the player having to be really precise about hitting the coin hit box. Various small bugs have been fixed as well.