Added in volcano
I modeled the volcano and exported it into Unity. However, I encountered an issue where only the outside of the volcano was visible, while the inside, which is crucial for the game, was not. Upon checking in Blender, I discovered that the normals were only facing outward. To resolve this, I added a solidify modifier, which allowed the normals to face inward. After making this adjustment, I re-exported the model back into Unity, and it worked as intended.