Can Continue Hover After Button Has Been Released
Changed the hover mechanic so when the player releases the hover button, they now stop performing the hover but can continue on with the time they have remaining if they hold the button again. The timer is then reset regardless of what its value is every time the player lands.
Also fixed a couple bugs with the coins dropping out of the palm trees.