Level Check Point System
Check points can now be placed around the map to alter where the player respawns when they die. The size of the collision can be adjusted to ensure the player makes contact with the checkpoint.
2 fields for an animator and trigger text can be assigned in the inspector to trigger an object in the level to react when the player activates the checkpoint.
The players position and rotation can be setup in the scene for where they will be respawned for a specific checkpoint, and the "Level Check Point" component in the inspector panel can be right clicked where the option "Set Spawn Position To Players Current Position" at the bottom of the drop down list can be selected to automatically set up the respawn position and rotation.