Grapple Point Bug Fixes
The player can no longer connect to grapple points through walls and platforms. Grapple points must be visible from the point of the player, and each one can now have a custom individual range specified to allow the player to detect them from different distances depending on the design of the level.