Potential Ledge Climb Up Mechanic (Currently Inactive)
Added the foundation for a mechanic where the player will hop up onto a platform if the character is almost able to reach the surface of it but they're not quite high enough. This has been de-activated for the time being though until it is decided this that this mechanic is required (and that there is enough time to continue developing and testing it further).