Menu / UI Activation Zones
Added triggers that activate UI to appear on screen. An alternate version allows UI to appear when a button is pressed while the player is within the zone. There is currently an issue with integrating this with the shop menu that will be resolved in the future.
Also addressed several bug fixes.