Multi Activator Puzzles
A new type of puzzle can now be made where activating multiple activators (such as levers, buttons or slingshot targets) will trigger the desired actions. To create one, the "Multi Object Puzzle" component can be added to a game object to define the target activators of the puzzle as well as the events that are called in response to all activators being triggered.